refactor: :fire: removed unnecessary login checksn and methods along with fixing the local dev fix
