+++ src/botPage/view/TradeInfoPanel/TradeTable.js
@@ -6,8 +6,8 @@ import { Table, Column } from 'react-virtualized';
 import PropTypes from 'prop-types';
 import { api_base } from '@api-base';
 import { translate } from '@i18n';
+import { isNumber, saveAs, appendRow, updateRow } from '@utils';
 import { observer as globalObserver } from '../../../common/utils/observer';
-import { appendRow, updateRow, saveAs, isNumber } from '../shared';
 import { roundBalance } from '../../common/tools';
 import * as style from '../style';
