+++ src/botPage/view/blockly/utils.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line max-classes-per-file
 import { translate } from '@i18n';
+import { saveAs } from '@utils';
 import { fieldGeneratorMapping } from './blocks/shared';
-import { saveAs } from '../shared';
 import config from '../../common/const';
 import { observer as globalObserver } from '../../../common/utils/observer';
 import { TrackJSError } from '../logger';
