+++ src/botPage/view/deriv/login-check.js
@@ -1,5 +1,4 @@
 import { parseQueryString, queryToObjectArray } from '@utils';
-import { AppConstants } from '@constants';
 import { getTokenList, setStorage, getStorage } from '@storage';
 import { addTokenIfValid } from '../../../common/appId';
 
@@ -10,24 +9,14 @@ export default function loginCheck() {
         if (!getTokenList().length) {
             if (tokenObjectList.length) {
                 addTokenIfValid(tokenObjectList[0].token, tokenObjectList).then(() => {
-                    const accounts = getTokenList();
-                    if (accounts.length) {
-                        setStorage(AppConstants.STORAGE_ACTIVE_TOKEN, accounts[0].token);
-                    }
                     window.location.replace(window.location.href.split('/?')[0]);
                     resolve();
                 });
             }
-            const active_account = getStorage('active_loginid') || '';
+
             let token_list = [];
             if (getStorage('client.accounts')) {
                 token_list = JSON.parse(getStorage('client.accounts'));
-            }
-            if (active_account && token_list.length) {
-                const active_token = token_list.find(account => account.accountName === active_account).token;
-                setStorage('activeToken', active_token);
-                window.location.reload();
-                resolve();
             }
             setStorage('tokenList', JSON.stringify(token_list));
         }
