+++ src/common/hooks/useLogout.js
@@ -1,5 +1,4 @@
 import { useDispatch } from 'react-redux';
-import { AppConstants } from '@constants';
 import { setStorage, syncWithDerivApp, updateTokenList } from '@storage';
 import { resetClient } from '../../botPage/view/deriv/store/client-slice';
 import { logoutAllTokens } from '../appId';
@@ -10,7 +9,6 @@ const useLogout = () => {
     const logout = () => {
         logoutAllTokens().then(() => {
             updateTokenList();
-            setStorage(AppConstants.STORAGE_ACTIVE_TOKEN, '');
             setStorage('active_loginid', null);
             syncWithDerivApp();
             dispatch(resetClient());
